requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
selenium==4.17.2
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0